chore(mobile): Phase 4 Step 36 – migrate auth screens to features/auth; keep legacy screen paths as re-exports\ndocs: update URBAN_REALTY_REFACTORING_MASTER_REPORT and Phase 4 log
